Replace visibleContent by contentDisplay in demo pages
