+++ pages/table/shift-selection.page.tsx
@@ -15,9 +15,11 @@ import {
   getMatchesCountText,
   paginationLabels,
   pageSizeOptions,
-  visibleContentOptions,
+  contentDisplayPreference,
+  defaultPreferences,
 } from './shared-configs';
 import ScreenshotArea from '../utils/screenshot-area';
+import { contentDisplayPreferenceI18nStrings } from '../common/i18n-strings';
 
 const allItems = generateItems();
 const ariaLabels: TableProps<Instance>['ariaLabels'] = {
@@ -28,11 +30,7 @@ const ariaLabels: TableProps<Instance>['ariaLabels'] = {
 };
 
 export default function App() {
-  const [preferences, setPreferences] = useState<CollectionPreferencesProps.Preferences>({
-    pageSize: 20,
-    visibleContent: ['id', 'type', 'dnsName', 'state'],
-    wrapLines: false,
-  });
+  const [preferences, setPreferences] = useState<CollectionPreferencesProps.Preferences>(defaultPreferences);
   const { items, actions, filteredItemsCount, collectionProps, filterProps, paginationProps } = useCollection(
     allItems,
     {
@@ -77,7 +75,7 @@ export default function App() {
             filteringAriaLabel="Filter instances"
           />
         }
-        visibleColumns={preferences.visibleContent}
+        columnDisplay={preferences.contentDisplay}
         selectedItems={selectedItems}
         selectionType={'multi'}
         trackBy={'id'}
@@ -94,9 +92,9 @@ export default function App() {
               title: 'Select page size',
               options: pageSizeOptions,
             }}
-            visibleContentPreference={{
-              title: 'Select visible columns',
-              options: visibleContentOptions,
+            contentDisplayPreference={{
+              ...contentDisplayPreference,
+              ...contentDisplayPreferenceI18nStrings,
             }}
             wrapLinesPreference={{
               label: 'Wrap lines',
