+++ pages/table/striped-rows.page.tsx
@@ -16,8 +16,10 @@ import {
   getMatchesCountText,
   paginationLabels,
   pageSizeOptions,
-  visibleContentOptions,
+  contentDisplayPreference,
+  defaultPreferences,
 } from './shared-configs';
+import { contentDisplayPreferenceI18nStrings } from '../common/i18n-strings';
 
 const allItems = generateItems();
 const ariaLabels: TableProps<Instance>['ariaLabels'] = {
@@ -29,10 +31,7 @@ const ariaLabels: TableProps<Instance>['ariaLabels'] = {
 
 export default function App() {
   const [preferences, setPreferences] = useState<CollectionPreferencesProps.Preferences>({
-    pageSize: 20,
-    visibleContent: ['id', 'type', 'dnsName', 'state'],
-    wrapLines: false,
-
+    ...defaultPreferences,
     // set to true for default striped rows.
     stripedRows: true,
   });
@@ -87,7 +86,7 @@ export default function App() {
             filteringAriaLabel="Filter instances"
           />
         }
-        visibleColumns={preferences.visibleContent}
+        columnDisplay={preferences.contentDisplay}
         preferences={
           <CollectionPreferences
             title="Preferences"
@@ -99,9 +98,9 @@ export default function App() {
               title: 'Select page size',
               options: pageSizeOptions,
             }}
-            visibleContentPreference={{
-              title: 'Select visible columns',
-              options: visibleContentOptions,
+            contentDisplayPreference={{
+              ...contentDisplayPreference,
+              ...contentDisplayPreferenceI18nStrings,
             }}
             wrapLinesPreference={{
               label: 'Wrap lines',
